requirements.txt
azure-ai-projects
azure-identity
streamlit
python-dotenv
azure-ai-inference